Handle empty string in client site name
parse_disco_string() returns an empty string if there is no client site
name in the NetLogon reply. Later in the code only a NULL check is used
to check for a missing client site name. To make sure the empty string
is not used as client site name it should be replaced with NULL.

https://bugs.freedesktop.org/show_bug.cgi?id=98143