Add add-service-principal and remove-service-principal options
Currently it is only possible to specific a service name for service
principals but not to set the full service principal. This is e.g.
needed if there is a service running on a host which should be reachable
by a different DNS name as well.

With this patch service principal can be added and removed by specifying
the full name.

Related to https://bugzilla.redhat.com/show_bug.cgi?id=1547014