library: use getaddrinfo with AI_CANONNAME to find a FQDN
Currently adcli creates service principals only with a short name if the
hostname of the client is a short name. This would fail is
Kerberos/GSSAPI clients will use the fully-qualified domain name (FQDN)
to access the host.

With this patch adcli tries to expand the short name by calling
getaddrinfo with the AI_CANONNAME hint.

Related to https://gitlab.freedesktop.org/realmd/adcli/issues/1