adenroll: add adcli_enroll_get_permitted_keytab_enctypes with tests
The new call does not only return the current encryption types set in AD
or a default list but filters them with the list of permitted encryption
types on the client. This makes sure the client can create and use the
keys.

Related to https://gitlab.freedesktop.org/realmd/adcli/issues/3